moved the secrets mock http client into its own file
