+++ secrets/main_test.go
@@ -1,51 +1,12 @@
 package secrets
 
 import (
-	"bytes"
-	"fmt"
-	"github.com/blend/go-sdk/assert"
-	"io/ioutil"
-	"net/http"
-	"net/url"
 	"testing"
+
+	"github.com/blend/go-sdk/assert"
 )
 
 // TestMain is the testing entrypoint.
 func TestMain(m *testing.M) {
 	assert.Main(m)
-}
-
-// NewMockHTTPClient returns a new mock http client.
-func NewMockHTTPClient() *MockHTTPClient {
-	return &MockHTTPClient{
-		contents: make(map[string]*http.Response),
-	}
-}
-
-// MockHTTPClient is a mock http client.
-type MockHTTPClient struct {
-	contents map[string]*http.Response
-}
-
-// With adds a mocked endpoint.
-func (mh *MockHTTPClient) With(verb string, url *url.URL, response *http.Response) *MockHTTPClient {
-	mh.contents[fmt.Sprintf("%s_%s", verb, url.String())] = response
-	return mh
-}
-
-// WithString adds a mocked endpoint.
-func (mh *MockHTTPClient) WithString(verb string, url *url.URL, contents string) *MockHTTPClient {
-	mh.contents[fmt.Sprintf("%s_%s", verb, url.String())] = &http.Response{
-		StatusCode: http.StatusOK,
-		Body:       ioutil.NopCloser(bytes.NewBuffer([]byte(contents))),
-	}
-	return mh
-}
-
-// Do implements HTTPClient.
-func (mh *MockHTTPClient) Do(req *http.Request) (*http.Response, error) {
-	if res, hasRes := mh.contents[fmt.Sprintf("%s_%s", req.Method, req.URL.String())]; hasRes {
-		return res, nil
-	}
-	return nil, fmt.Errorf("not found")
 }
